moves the logic to update the rounds querystring into the helper function file
